(feat) Apply some design changes
