+++ src/components/empty-state/empty-state.scss
@@ -1,6 +1,6 @@
+@use '@carbon/styles/scss/colors';
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
 
 .action {
   margin-bottom: spacing.$spacing-03;
@@ -8,7 +8,7 @@
 
 .content {
   @include type.type-style("heading-compact-01");
-  color: $text-02;
+  color: colors.$gray-20;
   margin-top: spacing.$spacing-05;
   margin-bottom: spacing.$spacing-03;
 }
@@ -16,14 +16,14 @@
 .desktopHeading {
   h4 {
     @include type.type-style('heading-compact-02');
-    color: $text-02;
+    color: colors.$gray-20;
   }
 }
 
 .tabletHeading {
   h4 {
     @include type.type-style('heading-03');
-    color: $text-02;
+    color: colors.$gray-20;
   }
 }
 
@@ -51,5 +51,5 @@
 
 .tile {
   text-align: center;
-  border: 1px solid $ui-03;
+  border: 1px solid colors.$gray-20;
 }
