+++ src/components/form-renderer/form-renderer.scss
@@ -1,10 +1,9 @@
 @use '@carbon/styles/scss/colors';
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
 
 .container {
-  background-color: white;
+  background-color: colors.$white-0;
   padding: 1rem;
   overflow-y: scroll;
   height: 100vh;
@@ -12,7 +11,7 @@
 
 .loadingContainer {
   @extend .container;
-  background-color: $ui-01;
+  background-color: colors.$gray-10;
   margin-top: 5rem;
 }
