+++ src/components/schema-editor/schema-editor.component.tsx
@@ -1,164 +1,26 @@
-import React, { useCallback, useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
+import React from "react";
 import AceEditor from "react-ace";
 import "ace-builds/webpack-resolver";
 import "ace-builds/src-noconflict/ext-language_tools";
-import { Button, InlineLoading } from "@carbon/react";
 import { useTranslation } from "react-i18next";
-import { OHRIFormSchema } from "@openmrs/openmrs-form-engine-lib";
-
-import { RouteParams, Schema } from "../../types";
 import styles from "./schema-editor.scss";
 
 type SchemaEditorProps = {
   isLoading: boolean;
-  onSchemaChange: (schema: Schema) => void;
-  schema: Schema;
+  invalidJsonErrorMessage: string;
+  onSchemaChange: (stringifiedSchema: string) => void;
+  stringifiedSchema: string;
 };
 
 const SchemaEditor: React.FC<SchemaEditorProps> = ({
-  isLoading,
+  invalidJsonErrorMessage,
   onSchemaChange,
-  schema,
+  stringifiedSchema,
 }) => {
   const { t } = useTranslation();
-  const { formUuid } = useParams<RouteParams>();
-  const isNewSchema = !formUuid;
-
-  const [stringifiedSchema, setStringifiedSchema] = useState(
-    schema ? JSON.stringify(schema, null, 2) : ""
-  );
-  const [invalidJsonErrorMessage, setInvalidJsonErrorMessage] = useState("");
-  const [isRendering, setIsRendering] = useState(false);
-
-  const resetErrorMessage = useCallback(() => {
-    setInvalidJsonErrorMessage("");
-  }, []);
-
-  const handleSchemaChange = useCallback((updatedSchema: string) => {
-    setStringifiedSchema(updatedSchema);
-  }, []);
-
-  const inputDummySchema = useCallback(() => {
-    const dummySchema: OHRIFormSchema = {
-      encounterType: "",
-      name: "Sample Form",
-      processor: "EncounterFormProcessor",
-      referencedForms: [],
-      uuid: "",
-      version: "1.0",
-      pages: [
-        {
-          label: "First Page",
-          sections: [
-            {
-              label: "A Section",
-              isExpanded: "true",
-              questions: [
-                {
-                  label: "A Question of type obs that renders a text input",
-                  type: "obs",
-                  questionOptions: {
-                    rendering: "text",
-                    concept: "a-system-defined-concept-uuid",
-                  },
-                  id: "sampleQuestion",
-                },
-              ],
-            },
-            {
-              label: "Another Section",
-              isExpanded: "true",
-              questions: [
-                {
-                  label:
-                    "Another Question of type obs whose answers get rendered as radio inputs",
-                  type: "obs",
-                  questionOptions: {
-                    rendering: "radio",
-                    concept: "system-defined-concept-uuid",
-                    answers: [
-                      {
-                        concept: "another-system-defined-concept-uuid",
-                        label: "Choice 1",
-                        conceptMappings: [],
-                      },
-                      {
-                        concept: "yet-another-system-defined-concept-uuid",
-                        label: "Choice 2",
-                        conceptMappings: [],
-                      },
-                      {
-                        concept: "yet-one-more-system-defined-concept-uuid",
-                        label: "Choice 3",
-                        conceptMappings: [],
-                      },
-                    ],
-                  },
-                  id: "anotherSampleQuestion",
-                },
-              ],
-            },
-          ],
-        },
-      ],
-    };
-
-    setStringifiedSchema(JSON.stringify(dummySchema, null, 2));
-    onSchemaChange({ ...dummySchema });
-  }, [onSchemaChange]);
-
-  const renderSchemaChanges = useCallback(() => {
-    setIsRendering(true);
-    resetErrorMessage();
-
-    try {
-      const parsedJson: Schema = JSON.parse(stringifiedSchema);
-      onSchemaChange(parsedJson);
-      setStringifiedSchema(JSON.stringify(parsedJson, null, 2));
-    } catch (error) {
-      setInvalidJsonErrorMessage(error.message);
-    }
-    setIsRendering(false);
-  }, [stringifiedSchema, onSchemaChange, resetErrorMessage]);
-
-  useEffect(() => {
-    setStringifiedSchema(JSON.stringify(schema, null, 2));
-  }, [schema]);
 
   return (
     <>
-      <div className={styles.actionButtons}>
-        {isLoading ? (
-          <InlineLoading
-            description={t("loadingSchema", "Loading schema") + "..."}
-          />
-        ) : null}
-
-        {isNewSchema && !stringifiedSchema ? (
-          <Button kind="secondary" onClick={inputDummySchema}>
-            {t("inputDummySchema", "Input dummy schema")}
-          </Button>
-        ) : null}
-
-        {schema ? (
-          <Button
-            disabled={isRendering}
-            kind="primary"
-            onClick={renderSchemaChanges}
-          >
-            {isRendering ? (
-              <InlineLoading
-                className={styles.spinner}
-                description={t("render", "Render" + "...")}
-              />
-            ) : (
-              <span>{t("renderChanges", "Render changes")}</span>
-            )}
-          </Button>
-        ) : null}
-      </div>
-
       {invalidJsonErrorMessage ? (
         <div className={styles.errorContainer}>
           <p className={styles.heading}>
@@ -173,7 +35,7 @@ const SchemaEditor: React.FC<SchemaEditorProps> = ({
         mode="json"
         theme="textmate"
         name="schemaEditor"
-        onChange={handleSchemaChange}
+        onChange={onSchemaChange}
         fontSize={15}
         showPrintMargin={false}
         showGutter={true}
